refactor(ui): extract feed/mount.ts — render-boundary helpers + ProvenanceChip
